PSYNEU-36 chore: Remove deprecated folder
