--- src/components/views/visualiseView/main/charts/mock.js
@@ -1,54 +1,0 @@
-export const mock = [
-  {
-    id: 'Company A',
-    name: 'Company A',
-    info: [
-      {
-        x: 1,
-        y: 10,
-        text: 'A',
-      },
-      {
-        x: 2,
-        y: 15,
-        text: 'B',
-      },
-      {
-        x: 3,
-        y: 13,
-        text: 'C',
-      },
-      {
-        x: 4,
-        y: 17,
-        text: 'D',
-      },
-    ],
-  },
-  {
-    id: 'Company B',
-    name: 'Company B',
-    info: [
-      {
-        x: 1,
-        y: 16,
-        text: 'W',
-      },
-      {
-        x: 2,
-        y: 5,
-        text: 'X',
-      },
-      {
-        x: 3,
-        y: 11,
-        text: 'Y',
-      },
-      {
-        x: 4,
-        y: 9,
-        text: 'Z',
-      },
-    ],
-  },
-];
